mcp_server/requirements.txt
fastapi
uvicorn
requests
fastmcp
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
dropbox
python-docx